clippy & fmt

+++ polkadot/node/subsystem-bench/src/lib/mock/prospective_parachains.rs
@@ -27,28 +27,28 @@ use std::collections::HashMap;
 use crate::availability::TestState;
 
 pub struct MockProspectiveParachains {
-    // Map relay parent -> candidate hashes (in order) used by the bench
-    per_relay_parent_candidates: HashMap<Hash, Vec<polkadot_primitives::CandidateHash>>,
-    state: Option<TestState>,
+	// Map relay parent -> candidate hashes (in order) used by the bench
+	per_relay_parent_candidates: HashMap<Hash, Vec<polkadot_primitives::CandidateHash>>,
+	state: Option<TestState>,
 }
 
 impl MockProspectiveParachains {
-    pub fn new() -> Self {
-        Self { per_relay_parent_candidates: HashMap::new(), state: None }
-    }
+	pub fn new() -> Self {
+		Self { per_relay_parent_candidates: HashMap::new(), state: None }
+	}
 
-    // Initialize from TestState so we can answer GetBackableCandidates
-    pub fn with_test_state(mut self, state: &TestState) -> Self {
-        for (relay_hash, receipts) in state.candidate_receipts.iter() {
-            let hashes = receipts
-                .iter()
-                .map(|r| -> polkadot_primitives::CandidateHash { r.hash() })
-                .collect::<Vec<_>>();
-            self.per_relay_parent_candidates.insert(*relay_hash, hashes);
-        }
-        self.state = Some(state.clone());
-        self
-    }
+	// Initialize from TestState so we can answer GetBackableCandidates
+	pub fn with_test_state(mut self, state: &TestState) -> Self {
+		for (relay_hash, receipts) in state.candidate_receipts.iter() {
+			let hashes = receipts
+				.iter()
+				.map(|r| -> polkadot_primitives::CandidateHash { r.hash() })
+				.collect::<Vec<_>>();
+			self.per_relay_parent_candidates.insert(*relay_hash, hashes);
+		}
+		self.state = Some(state.clone());
+		self
+	}
 }
 
 #[overseer::subsystem(ProspectiveParachains, error=SubsystemError, prefix=self::overseer)]
@@ -74,24 +74,38 @@ impl MockProspectiveParachains {
 					ProspectiveParachainsMessage::GetMinimumRelayParents(_relay_parent, tx) => {
 						tx.send(vec![]).unwrap();
 					},
-                    ProspectiveParachainsMessage::GetBackableCandidates(relay_parent, para_id, count, _ancestors, tx) => {
-                        // Return up to `count` candidate hashes for the NEXT block's candidates of this para.
-                        let mut response = Vec::new();
-                        if let Some(state) = &self.state {
-                            // find index of current relay_parent in block_infos
-                            if let Some((idx, _)) = state.block_infos.iter().enumerate().find(|(_, bi)| bi.hash == relay_parent) {
-                                let next_idx = ((idx + 1) % state.block_infos.len()) as usize;
-                                let next_hash = state.block_infos[next_idx].hash;
-                                if let Some(receipts) = state.candidate_receipts.get(&next_hash) {
-                                    for r in receipts.iter() {
-                                        if r.descriptor.para_id() == para_id && response.len() < count as usize {
-                                            response.push((r.hash(), relay_parent));
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        let _ = tx.send(response);
+					ProspectiveParachainsMessage::GetBackableCandidates(
+						relay_parent,
+						para_id,
+						count,
+						_ancestors,
+						tx,
+					) => {
+						// Return up to `count` candidate hashes for the NEXT block's candidates of
+						// this para.
+						let mut response = Vec::new();
+						if let Some(state) = &self.state {
+							// find index of current relay_parent in block_infos
+							if let Some((idx, _)) = state
+								.block_infos
+								.iter()
+								.enumerate()
+								.find(|(_, bi)| bi.hash == relay_parent)
+							{
+								let next_idx = (idx + 1) % state.block_infos.len();
+								let next_hash = state.block_infos[next_idx].hash;
+								if let Some(receipts) = state.candidate_receipts.get(&next_hash) {
+									for r in receipts.iter() {
+										if r.descriptor.para_id() == para_id &&
+											response.len() < count as usize
+										{
+											response.push((r.hash(), relay_parent));
+										}
+									}
+								}
+							}
+						}
+						let _ = tx.send(response);
 					},
 					ProspectiveParachainsMessage::GetHypotheticalMembership(req, tx) => {
 						tx.send(
